requirements.txt
absl-py==1.3.0
altgraph==0.17.3
asttokens==2.2.1
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
blinker==1.4
cachetools==5.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cloud-init==22.4.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.6
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.3.0
easydict==1.10
entrypoints==0.3
executing==1.2.0
filelock==3.8.2
Flask==2.2.2
fonttools==4.38.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
httplib2==0.14.0
hyperlink==19.0.0
idna==3.4
importlib-metadata==5.1.0
incremental==16.10.1
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.2
parso==0.8.3
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycocotools==2.0.6
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.3.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.6
PyYAML==5.3.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
ssh-import-id==5.10
stack-data==0.6.2
stevedore==4.1.1
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
testresources==2.0.0
thop==0.1.1.post2209072238
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.6.0
Twisted==18.9.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.12
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.26.14
virtualenv==20.17.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.2.2
wifite==2.5.2
zipp==3.12.0
zope.interface==4.7.1
